Using Functions to Implement a Caesar Cipher
We wrote a program that implements a Caesar cipher, which is a simple method of encryption.

A Caesar cipher takes the letters of a message and shifts each letter along the alphabet by a certain number of places.

In this lab,

We created user-defined functions
Use several functions to implement a Caesar cipher encryption program